tests/wpt/web-platform-tests/css/css-scroll-snap/parsing/scroll-margin-block-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-margin-block with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-margin-block">
<meta name="assert" content="scroll-margin-block supports the full grammar '<length>{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("scroll-margin-block-start", "-10px");
test_valid_value("scroll-margin-block-start", "calc(2em + 3ex)");


test_valid_value("scroll-margin-block-end", "-10px");
test_valid_value("scroll-margin-block-end", "calc(2em + 3ex)");


test_valid_value("scroll-margin-block", "-10px");
test_valid_value("scroll-margin-block", "calc(2em + 3ex)");

test_valid_value("scroll-margin-block", "1px 2px");
</script>
</body>
</html>
